requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
bleak>=0.14.3
pytest>=6.2.5
pytest-asyncio>=0.18.3
httpx>=0.23.0
python-dotenv>=0.19.1